# LoginAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LoginAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Login": {
      "post": {
        "tags": [
          "Login"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginInputModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        }
      }
    },
    "/api/Usuario/RegistrarUsuario": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioInputModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UsuarioViewModel"
              }
            }
          }
        }
      }
    },
    "/api/Usuario/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UsuarioViewModel"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/ActualizarClave": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ActualizarClaveInputModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "ActualizarClaveInputModel": {
      "required": [
        "claveActual",
        "confirmacionClave",
        "correo",
        "nuevaClave"
      ],
      "type": "object",
      "properties": {
        "correo": {
          "pattern": "^\\w+([-+.']\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*$",
          "type": "string"
        },
        "claveActual": {
          "type": "string"
        },
        "nuevaClave": {
          "type": "string"
        },
        "confirmacionClave": {
          "type": "string"
        }
      }
    },
    "LoginInputModel": {
      "required": [
        "clave",
        "correo"
      ],
      "type": "object",
      "properties": {
        "correo": {
          "pattern": "^\\w+([-+.']\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*$",
          "type": "string"
        },
        "clave": {
          "type": "string"
        }
      }
    },
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "correo": {
          "type": "string"
        },
        "clave": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "rol": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UsuarioInputModel": {
      "required": [
        "clave",
        "confirmacionClave",
        "correo",
        "nombre"
      ],
      "type": "object",
      "properties": {
        "correo": {
          "pattern": "^\\w+([-+.']\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*$",
          "type": "string"
        },
        "clave": {
          "maxLength": 50,
          "minLength": 6,
          "type": "string"
        },
        "confirmacionClave": {
          "maxLength": 50,
          "minLength": 6,
          "type": "string"
        },
        "nombre": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "apellido": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      }
    },
    "UsuarioViewModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "correo": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "rol": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}